+++ api/globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:62388/"}
+{"mongoUri":"mongodb://127.0.0.1:49476/"}
