+++ api/src/models/show/show.model.js
@@ -1,55 +1,165 @@
-const mongoose = require("mongoose");
-const idvalidator = require("mongoose-id-validator");
-const { SHOW_GENRES, COUNTRIES } = require("../enum");
+const { default: mongoose } = require("mongoose");
+const {
+  DEFAULT_PAGE_SIZE,
+  showSortOptions,
+  customProjection,
+} = require("../../configs/route.config");
+const Show = require("./Show");
 
-const episodeSchema = new mongoose.Schema(
-  {
-    title: { type: String, required: true },
-    episodeNumber: { type: Number },
-    airDate: { type: Date },
-    runtime: { type: Number },
-    overview: { type: String },
-    thumbnailUrl: { type: String },
-    videoUrl: { type: String },
-  },
-  { timestamps: true }
-);
+async function exists(title) {
+  return (await Show.exists({ title: title })) !== null;
+}
 
-const seasonSchema = new mongoose.Schema(
-  {
-    title: { type: String, required: true },
-    seasonNumber: { type: Number },
-    overview: { type: String },
-    releaseDate: { type: Date },
-    posterUrl: { type: String },
-    backdropUrl: { type: String },
-    episodes: { type: [episodeSchema] },
-  },
-  { timestamps: true }
-);
+// Add
+function addShow(show) {
+  return Show.create(show);
+}
 
-const showSchema = new mongoose.Schema(
-  {
-    title: { type: String, required: true, unique: true },
-    tagLine: { type: String },
-    overview: { type: String },
-    adult: { type: Boolean },
-    imdbID: { type: String },
-    genres: { type: [String], enum: SHOW_GENRES },
-    countries: { type: [String], enum: COUNTRIES },
-    cast: { type: [mongoose.Schema.Types.ObjectId], ref: "Person" },
-    directors: { type: [mongoose.Schema.Types.ObjectId], ref: "Person" },
-    trailers: { type: [String] },
-    posterUrl: { type: String },
-    thumbnailUrl: { type: String },
-    backdropUrl: { type: String },
-    firstAirDate: { type: Date },
-    lastAirDate: { type: Date },
-    seasons: { type: [seasonSchema] },
-  },
-  {
-    timestamps: true,
+// Get Shows
+function getAllShows() {
+  return Show.find();
+}
+
+async function getPaginatedShows(
+  filter = null,
+  sort = null,
+  page = 1,
+  project = customProjection.ITEM_BASE_INFO
+) {
+  try {
+    const [result] = await Show.aggregate([
+      { $match: filter },
+      {
+        $facet: {
+          docs: [
+            { $sort: sort },
+            { $skip: DEFAULT_PAGE_SIZE * (page - 1) },
+            { $limit: DEFAULT_PAGE_SIZE },
+            { $project: project },
+          ],
+          meta: [{ $count: "total_documents" }],
+        },
+      },
+      { $unwind: "$meta" },
+    ]);
+
+    const totalDocs = result?.meta.total_documents || 0;
+    const output = {
+      docs: result?.docs || [],
+      page: page,
+      pageSize: DEFAULT_PAGE_SIZE,
+      total_pages: Math.ceil(totalDocs / DEFAULT_PAGE_SIZE),
+      total_documents: totalDocs,
+    };
+
+    return output;
+  } catch (error) {
+    throw error;
   }
-);
+}
 
-module.exports = mongoose.model("Show", showSchema);
+async function getShows({
+  genre = null,
+  country = null,
+  year = null,
+  sort = showSortOptions.lastAirDate,
+  page = 1,
+}) {
+  const filter = {};
+  if (genre) filter.genres = { $all: [genre] };
+  if (country) filter.countries = { $all: [country] };
+  if (year)
+    filter.lastAirDate = {
+      $gte: new Date("2021-01-01"),
+      $lte: new Date("2021-12-31"),
+    };
+
+  try {
+    return await getPaginatedShows(filter, sort, page);
+  } catch (error) {
+    throw error;
+  }
+}
+
+async function getShowsByTitle(query, page = 1) {
+  try {
+    const filter = { $text: { $search: query } };
+    const sort = { score: { $meta: "textScore" } };
+    return await getPaginatedShows(filter, sort, page);
+  } catch (error) {
+    throw error;
+  }
+}
+
+async function getSimilarShows(id) {
+  try {
+    const { genres } = await Show.findById(id, { genres: 1 });
+    const projection = customProjection.ITEM_BASE_INFO;
+    projection.numSimilar = {
+      $size: { $setIntersection: [genres, "$genres"] },
+    };
+    const similarShows = await Show.aggregate([
+      {
+        $match: {
+          _id: { $ne: new mongoose.Types.ObjectId(id) },
+          genres: { $in: genres },
+        },
+      },
+      { $project: projection },
+      { $sort: { numSimilar: -1, lastAirDate: -1 } },
+      { $limit: DEFAULT_PAGE_SIZE },
+      { $project: { numSimilar: 0 } },
+    ]);
+    return similarShows;
+  } catch (error) {
+    throw error;
+  }
+}
+
+// Get Single Shows
+function getShowByID(id) {
+  return Show.findById(id, customProjection.ITEM_FULL_INFO)
+    .populate("cast", customProjection.PERSON_BRIEF_INFO)
+    .populate("directors", customProjection.PERSON_BRIEF_INFO);
+}
+
+function getShowByTitle(title) {
+  return Show.findOne({ title: title });
+}
+
+function getRandomShow() {
+  return Show.aggregate([
+    { $sample: { size: 1 } },
+    { $project: customProjection.ITEM_BASE_INFO },
+  ]);
+}
+
+// Update
+function updateShow(id, updateData) {
+  return Show.findByIdAndUpdate(id, updateData, {
+    returnDocument: "after",
+    runValidators: true,
+    projection: customProjection.ITEM_FULL_INFO,
+  });
+}
+
+// Delete
+function deleteShowByID(id) {
+  return Show.findByIdAndDelete(id, {
+    projection: customProjection.ITEM_FULL_INFO,
+  });
+}
+
+module.exports = {
+  exists,
+  addShow,
+  getAllShows,
+  getShows,
+  getShowsByTitle,
+  getSimilarShows,
+  getShowByID,
+  getShowByTitle,
+  getRandomShow,
+  updateShow,
+  deleteShowByID,
+};
