+++ api/src/models/user/User.js
@@ -1,0 +1,15 @@
+const mongoose = require("mongoose");
+
+const userSchema = new mongoose.Schema(
+  {
+    username: { type: String, required: true, unique: true },
+    email: { type: String, required: true, unique: true },
+    password: { type: String, required: true },
+    profileImage: { type: String },
+    isAdmin: { type: Boolean, default: false },
+    refreshToken: { type: String },
+  },
+  { timestamps: true }
+);
+
+module.exports = mongoose.model("User", userSchema);
